Re-add validations for labels value to be not null
Previously there is a false positive match if the input of the label filter is a key that does not exist in DB and value is null. Accessing `map[non-existent-key]` in golang returns a null value, thus causing `reflect.DeepEqual` to become true, and add the subdomain to the end result.